Make changes in order to use async version, fix flake8 errors.

+++ City_CRUD_API/crud.py
@@ -1,59 +1,71 @@
-from sqlalchemy.orm import Session
+from sqlalchemy.future import select
+from sqlalchemy.ext.asyncio import AsyncSession
 
 from City_CRUD_API import models
 from City_CRUD_API.schemas import CityCreate
 
 
-def get_all_cities(db: Session):
-    return db.query(models.City).all()
+async def get_all_cities(db: AsyncSession):
+    result = await db.execute(select(models.City))
+    return result.scalars().all()
 
 
-def create_city(db: Session, city: CityCreate):
+async def create_city(db: AsyncSession, city: CityCreate):
     db_city = models.City(
         name=city.name,
         additional_info=city.additional_info,
     )
     db.add(db_city)
-    db.commit()
-    db.refresh(db_city)
+    await db.commit()
+    await db.refresh(db_city)
 
     return db_city
 
 
-def get_city_by_id(db: Session, city_id: int):
-    return (
-        db.query(models.City).
-        filter(models.City.id == city_id).first()
+async def get_city_by_id(db: AsyncSession, city_id: int):
+    result = await db.execute(
+        select(models.City).filter(models.City.id == city_id)
     )
+    return result.scalars().first()
 
 
-def get_city_by_name(db: Session, name: str):
+async def get_city_by_name(db: AsyncSession, name: str):
     """
     In case the same city already exists
     """
-    return (
-        db.query(models.City).filter(models.City.name == name).first()
+    result = await db.execute(
+        select(models.City).filter(models.City.name == name)
     )
+    return result.scalars().first()
 
 
-def update_city(
-        db: Session,
+async def update_city(
+        db: AsyncSession,
         city_id: int,
         name: str = None,
         additional_info: str = None,
 ):
-    city = get_city_by_id(db, city_id)
+    city = await get_city_by_id(db, city_id)
+    if not city:
+        return None
+
     if name:
         city.name = name
     if additional_info:
         city.additional_info = additional_info
-    db.commit()
-    db.refresh(city)
+
+    await db.commit()
+    await db.refresh(city)
+
     return city
 
 
-def delete_city(db: Session, city_id: int):
-    city = get_city_by_id(db, city_id)
-    db.delete(city)
-    db.commit()
+async def delete_city(db: AsyncSession, city_id: int):
+    city = await get_city_by_id(db, city_id)
+    if not city:
+        return {"detail": f"City with ID {city_id} not found."}
+
+    await db.delete(city)
+    await db.commit()
+
     return {"detail": f"City with ID {city_id} deleted successfully."}
